Announcement Commit
aipsetup and Horizon Linux are going to be continued.

It will be renamed back to Lilith Linux

this girl (https://vk.com/id9001832) has betrayed me no metter sacrifices
I've done for here. She appeared not worth it.. she betrayed my hopes.

So.. when?

I think soon. this year work will be continued.

Toolkit (that's right) is ganna be many improved.

First of all, this gonna be a new workbench, to perform first stage
of building. this is gonna be more separate process.

Donna use docker or some other VT tools for more vacuumic development of
first stage. If RH have something for this - I will use it (podman?).

As You, probably, noticed, Lilith is somewhat Fedorish.. this is gonna continued.
Bu it's not just fedorish, It's one-man project still.

Everything is nona be ok. Keep Your self alive! and listen to David Bowie and Freddie Mercury ;-)